Move flatten_module and mixed_precision to `transforms`.
Fix CI.
