+++ .github/workflows/ci.yml
@@ -34,5 +34,4 @@ jobs:
     - name: Test with pytype
       run: |
         pip install pytype
-        pytype
-        pytype tests
+        pytype pax tests
