+++ pax/__init__.py
@@ -4,9 +4,11 @@ from .module import Module, PaxFieldKind
 from .pax_transforms import grad, jit, pmap, vmap
 from .rng import next_rng_key, seed_rng_key
 from .transforms import (
+    apply_mixed_precision_policy,
     apply_updates,
     enable_eval_mode,
     enable_train_mode,
+    flatten_module,
     freeze_parameter,
     grad_with_aux,
     scan_bug,
