+++ pax/module.py
@@ -475,4 +475,3 @@ class Module:
 
     def __call__(self, *args, **kwargs):
         raise NotImplementedError
-
