+++ pax/nn/__init__.py
@@ -4,13 +4,11 @@ from .conv import Conv1D, Conv1DTranspose, Conv2D, Conv2DTranspose
 from .dropout import Dropout
 from .ema import EMA
 from .embed import Embed
-from .flatten import FlattenModule
 from .group_norm import GroupNorm
 from .identity import Identity
 from .lambda_module import Lambda
 from .layer_norm import LayerNorm
 from .linear import Linear
-from .mixed_precision import MixedPrecisionModule
 from .pool import avg_pool, max_pool
 from .recurrent import GRU, LSTM, GRUState, LSTMState
 from .rng_seq import RngSeq
