--- setup.cfg
@@ -1,2 +1,0 @@
-[pytype]
-inputs = pax
