+++ tests/test_flatten_module.py
@@ -1,11 +1,12 @@
+from typing import List
+
 import jax
 import pax
-import pytest
 
 
 def test_flatten_module():
     f = pax.nn.Linear(4, 4)
-    g = pax.nn.FlattenModule(f)
+    g = pax.flatten_module(f)
     k = pax.select_parameter(g)
     assert jax.tree_structure(g) == jax.tree_structure(k)
     h = g.update(k)
@@ -13,6 +14,8 @@ def test_flatten_module():
 
 def test_none_state():
     class M(pax.Module):
+        s: List
+
         def __init__(self):
             super().__init__()
             self.register_state_subtree("s", [])
@@ -30,29 +33,31 @@ def test_flatten_non_callable_module():
 
     m = M()
 
-    with pytest.raises(ValueError):
-        n = pax.nn.FlattenModule(m)
+    # with pytest.raises(ValueError):
+    n = pax.flatten_module(m)
+    k = n.unflatten()
+    assert type(k) == M
 
 
 def test_flatten_module_freeze():
     a = pax.nn.Linear(1, 1)
-    b = pax.nn.FlattenModule(a)
+    b = pax.flatten_module(a)
 
-    with pytest.raises(RuntimeError):
-        c = b.freeze()
+    # with pytest.raises(RuntimeError):
+    c = pax.freeze_parameter(b)
 
 
 def test_flatten_module_unfreeze():
     a = pax.nn.Linear(1, 1)
-    b = pax.nn.FlattenModule(a)
+    b = pax.flatten_module(a)
 
-    with pytest.raises(RuntimeError):
-        c = b.unfreeze()
+    # with pytest.raises(RuntimeError):
+    # c = b.unfreeze()
 
 
 def test_flatten_module_eval():
     a = pax.nn.Linear(1, 1)
-    b = pax.nn.FlattenModule(a)
+    b = pax.flatten_module(a)
 
     # with pytest.raises(RuntimeError):
     c = pax.enable_eval_mode(b)
