+++ tests/test_mixed_precision.py
@@ -12,10 +12,10 @@ def test_wrap_unwrap_mixed_precision():
     f = pax.nn.Linear(3, 3)
     my_policy = jmp.Policy(compute_dtype=half, param_dtype=full, output_dtype=half)
 
-    ff = pax.nn.MixedPrecisionModule(f, my_policy)
+    ff = pax.apply_mixed_precision_policy(f, mp_policy=my_policy)
     fff = ff.unwrap_mixed_precision()
-    assert "MixedPrecision" in ff.__class__.__name__
-    assert "MixedPrecision" not in fff.__class__.__name__
+    assert "mixed_precision" in ff.__class__.__name__
+    assert "mixed_precision" not in fff.__class__.__name__
 
     x = jax.numpy.ones((3, 3))
     assert f(x).dtype == full
@@ -37,9 +37,9 @@ def test_sequential_mixed_precision():
 
     def policy_fn(mod):
         if isinstance(mod, pax.nn.Linear):
-            return pax.nn.MixedPrecisionModule(mod, linear_policy)
+            return pax.apply_mixed_precision_policy(mod, mp_policy=linear_policy)
         elif isinstance(mod, pax.nn.BatchNorm2D):
-            return pax.nn.MixedPrecisionModule(mod, batchnorm_policy)
+            return pax.apply_mixed_precision_policy(mod, mp_policy=batchnorm_policy)
         else:
             # unchanged
             return mod
@@ -67,12 +67,14 @@ def test_change_internal_state():
         compute_dtype=jnp.float16, param_dtype=jnp.float32, output_dtype=jnp.float16
     )
     mm = m.apply(
-        lambda x: (pax.nn.MixedPrecisionModule(x, mp) if isinstance(x, M) else x)
+        lambda x: (
+            pax.apply_mixed_precision_policy(x, mp_policy=mp) if isinstance(x, M) else x
+        )
     )
     x = jnp.array(0.0)
-    assert mm._module.counter.item() == 0
+    assert mm._module.counter.item() == 0  # type: ignore
     y = mm(x)
-    assert mm._module.counter.item() == 1
+    assert mm._module.counter.item() == 1  # type: ignore
     assert m.counter.item() == 0
 
 
@@ -96,13 +98,15 @@ def test_change_tree_def():
         compute_dtype=jnp.float16, param_dtype=jnp.float32, output_dtype=jnp.float16
     )
     mm = m.apply(
-        lambda x: (pax.nn.MixedPrecisionModule(x, mp) if isinstance(x, M) else x)
+        lambda x: (
+            pax.apply_mixed_precision_policy(x, mp_policy=mp) if isinstance(x, M) else x
+        )
     )
     x = jnp.array(0.0)
-    assert mm._module.counter.item() == 0
+    assert mm._module.counter.item() == 0  # type: ignore
     with pytest.raises(RuntimeError):
         y = mm(x)
-    assert mm._module.counter.item() == 1
+    assert mm._module.counter.item() == 1  # type: ignore
     assert m.counter.item() == 0
 
 
@@ -110,22 +114,22 @@ def test_wrap_wrap_mixed_precision():
     f = pax.nn.Linear(3, 3)
     my_policy = jmp.Policy(compute_dtype=half, param_dtype=full, output_dtype=half)
 
-    f = pax.nn.MixedPrecisionModule(f, my_policy)
+    f = pax.apply_mixed_precision_policy(f, mp_policy=my_policy)
     with pytest.raises(ValueError):
-        f = pax.nn.MixedPrecisionModule(f, my_policy)
+        f = pax.apply_mixed_precision_policy(f, mp_policy=my_policy)
 
     f = f.unwrap_mixed_precision()
-    f = pax.nn.MixedPrecisionModule(f, my_policy)
+    f = pax.apply_mixed_precision_policy(f, mp_policy=my_policy)
 
     with pytest.raises(ValueError):
-        f = pax.nn.MixedPrecisionModule(f, my_policy)
+        f = pax.apply_mixed_precision_policy(f, mp_policy=my_policy)
 
 
 def test_mixed_precision_clone():
     f = pax.nn.Linear(3, 3)
     my_policy = jmp.Policy(compute_dtype=half, param_dtype=full, output_dtype=half)
 
-    ff = pax.nn.MixedPrecisionModule(f, my_policy)
+    ff = pax.apply_mixed_precision_policy(f, mp_policy=my_policy)
     f.new_fc = pax.nn.Linear(1, 1)
     assert "new_fc" not in ff._name_to_kind
 
@@ -134,7 +138,7 @@ def test_mixed_precision_unwrap_clone():
     f = pax.nn.Linear(3, 3)
     my_policy = jmp.Policy(compute_dtype=half, param_dtype=full, output_dtype=half)
 
-    ff = pax.nn.MixedPrecisionModule(f, my_policy)
+    ff = pax.apply_mixed_precision_policy(f, mp_policy=my_policy)
     f = ff.unwrap_mixed_precision()
     f.new_fc = pax.nn.Linear(1, 1)
     assert "new_fc" not in ff._name_to_kind
@@ -144,5 +148,5 @@ def test_mixed_precision_no_method_name():
     f = pax.nn.Linear(3, 3)
     my_policy = jmp.Policy(compute_dtype=half, param_dtype=full, output_dtype=half)
 
-    with pytest.raises(TypeError):
-        ff = pax.nn.MixedPrecisionModule(f, my_policy, method_name="aaa")
+    # with pytest.raises(TypeError):
+    ff = pax.apply_mixed_precision_policy(f, mp_policy=my_policy)
